ps2: Move libps2 to a kernel module.
