+++ config/hyper.cfg
@@ -5,7 +5,7 @@ protocol=ultra
 higher-half-exclusive=true
 kernel-as-module=true
 video-mode=auto
-cmdline='-initrd-module=initrd -initrd-driver-module=initrd_driver --root-fs-partid initrd --working-set-cap=8388608 --load-modules=/slowfat --initial-swap-size=67108864 --log-level=0 --no-init --no-smp --init-args /usr/bin/bash'
+cmdline='-initrd-module=initrd -initrd-driver-module=initrd_driver --root-fs-partid initrd --working-set-cap=8388608 --load-modules=/libps2,/slowfat --initial-swap-size=67108864 --log-level=0 --no-init --no-smp --init-args /usr/bin/bash'
 
 binary:
 	path="/obos/oboskrnl"
