+++ scripts/generate_initrd-x86_64.sh
@@ -11,6 +11,7 @@ cp out/slowfat tar
 cp out/bochs_vbe tar
 cp out/r8169 tar
 cp out/i8042 tar
+cp out/libps2 tar
 #cp out/init tar
 if [[ ! -d tar/dev ]]
 then
